# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Construction ERP API (MVP)
  version: "1.0.0"
  description: |
    Minimal API for Projects, Clients (with Contacts), Tasks, Variations, Documents, and AuditLog (read-only).
servers:
  - url: http://localhost:3001
paths:
  /api/health:
    get:
      summary: Health check
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                example: { ok: true }

  /api/clients:
    get:
      summary: List clients
      parameters:
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: page
          schema: { type: integer, minimum: 1 }
        - in: query
          name: pageSize
          schema: { type: integer, minimum: 1, maximum: 200 }
        - in: query
          name: sort
          schema: { type: string, default: name }
        - in: query
          name: order
          schema: { type: string, enum: [asc, desc], default: asc }
      responses:
        "200":
          description: Clients
          content:
            application/json:
              schema:
                type: object
                properties:
                  page: { type: integer }
                  pageSize: { type: integer }
                  total: { type: integer }
                  rows:
                    type: array
                    items: { $ref: "#/components/schemas/ClientListItem" }
    post:
      summary: Create client
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateClient" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Client" }

  /api/clients/{id}:
    get:
      summary: Get client by id (with contacts and projects summary)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Client
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ClientDetail" }
        "404":
          description: Not found
    put:
      summary: Update client
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateClient" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ClientDetail" }
        "404":
          description: Not found
    delete:
      summary: Soft delete client
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204": { description: No Content }
        "404": { description: Not found }

  /api/projects:
    get:
      summary: List projects
      parameters:
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100 }
        - in: query
          name: offset
          schema: { type: integer, minimum: 0 }
        - in: query
          name: sort
          schema: { type: string, default: startDate:desc }
        - in: query
          name: clientId
          schema: { type: integer }
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: type
          schema: { type: string }
      responses:
        "200":
          description: Projects
          content:
            application/json:
              schema:
                type: object
                properties:
                  total: { type: integer }
                  projects:
                    type: array
                    items: { $ref: "#/components/schemas/Project" }
    post:
      summary: Create project
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateProject" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }

  /api/projects/{id}:
    get:
      summary: Get project by id (with client, tasks, variations)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProjectDetail" }
        "404":
          description: Not found
    put:
      summary: Update project
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateProject" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
        "404":
          description: Not found
    delete:
      summary: Soft delete project
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204": { description: No Content }
        "404": { description: Not found }

  /api/tasks:
    get:
      summary: List tasks
      parameters:
        - in: query
          name: projectId
          schema: { type: integer }
        - in: query
          name: statusId
          schema: { type: integer }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100 }
        - in: query
          name: offset
          schema: { type: integer, minimum: 0 }
        - in: query
          name: sort
          schema: { type: string, default: dueDate:desc }
      responses:
        "200":
          description: Tasks
          content:
            application/json:
              schema:
                type: object
                properties:
                  total: { type: integer }
                  tasks:
                    type: array
                    items: { $ref: "#/components/schemas/Task" }
    post:
      summary: Create task
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateTask" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Task" }

components:
  schemas:
    ClientListItem:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        companyRegNo: { type: string }
        projectsCount: { type: integer }
    Client:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
    CreateClient:
      type: object
      required: [name]
      properties:
        name: { type: string }
        companyRegNo: { type: string }
        vatNo: { type: string }
        address1: { type: string }
        address2: { type: string }
        city: { type: string }
        county: { type: string }
        postcode: { type: string }
        primaryContact:
          type: object
          properties:
            firstName: { type: string }
            lastName: { type: string }
            email: { type: string }
            phone: { type: string }
            role: { type: string }
    UpdateClient:
      allOf:
        - $ref: "#/components/schemas/CreateClient"
    ClientDetail:
      allOf:
        - $ref: "#/components/schemas/Client"
        - type: object
          properties:
            contacts:
              type: array
              items:
                type: object
                properties:
                  id: { type: integer }
                  name: { type: string }
                  email: { type: string }
                  phone: { type: string }
            projects:
              type: array
              items:
                type: object
                properties:
                  id: { type: integer }
                  code: { type: string }
                  name: { type: string }
                  status: { type: string }
    Project:
      type: object
      properties:
        id: { type: integer }
        code: { type: string }
        name: { type: string }
        status:
          type: string
        client:
          type: object
          properties:
            id: { type: integer }
            name: { type: string }
    CreateProject:
      type: object
      required: [clientId, code, name]
      properties:
        clientId: { type: integer }
        code: { type: string }
        name: { type: string }
        statusId: { type: integer }
        typeId: { type: integer }
        description: { type: string }
        startDate: { type: string, format: date-time }
        endDate: { type: string, format: date-time }
    UpdateProject:
      allOf:
        - $ref: "#/components/schemas/CreateProject"
    ProjectDetail:
  /api/tasks/{id}:
    get:
      summary: Get task by id
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Task
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Task" }
        "404": { description: Not found }
    put:
      summary: Update task
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateTask" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Task" }
        "404": { description: Not found }
    delete:
      summary: Soft delete task
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204": { description: No Content }
        "404": { description: Not found }
      allOf:
        - $ref: "#/components/schemas/Project"
        - type: object
          properties:
            tasks:
              type: array
              items: { $ref: "#/components/schemas/Task" }
            variations:
              type: array
              items:
                type: object
                properties: {}
    Task:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: string }
        projectId: { type: integer }
        statusId: { type: integer }
        dueDate: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
    CreateTask:
      type: object
      required: [title, projectId, statusId]
      properties:
        title: { type: string }
        description: { type: string }
        projectId: { type: integer }
        statusId: { type: integer }
        dueDate: { type: string, format: date-time }
    UpdateTask:
      allOf:
        - $ref: "#/components/schemas/CreateTask"